Support non-US keyboard layout on macOS
This commit implements a look up table to map between CGKeyCode
on the physical keyboard and the actual output character from the
selected input source. This table is built once when the application
starts.

Currently, we have to restart the application if we change the
input source. This can be improved in the future.

Fixes #20