Support non-US keyboard layout on macOS - Part 2
Rebuilding the keyboard layout map when the UI is updated.
Normally, the UI is updated when the user changed the input
config state, like, toggle the Vietnamese input on/off or
other behaviors. This might be a good candidate to triggering
the rebuild, otherwise, we have to rebuild on every keystroke.