feat: Auto toggle Vietnamese input mode based on application name
Closes #38

With this change, we track the front-most application on every event (that might sound inefficient, but let's ship it and see how things goes).

The auto-toggling rule are implemented as:

- If the app is configured as Vietnamese, switch to Vietnamese mode
- If the app is configured as English, switch to English mode
- If the app is not configured, do nothing

Also, when the user toggling typing mode manually, we will track it and store the config.

All the configs are stored in ~/.goxkey file.